Restructuring: spliting restart button and typing test box on its own components to make it more modular
